refactor: 디바이스 API·타입 중복 제거 (#101)
온보딩과 설정이 GET /devices를 각자 정의해 쓰고 있었다. 온보딩 쪽 타입은
battery_level을 number로 선언했지만 서버는 null을 준다. 타입이 거짓말을 하고 있었다.

- onboarding/apis/deviceApi.ts, onboarding/types/deviceTypes.ts 삭제
- deviceApiTypes.ts와 getDevices.ts를 shared로 이동
  (온보딩이 pages/setting을 참조하는 역참조를 만들지 않기 위함)
- 온보딩 스토어의 connectedDevice 타입을 DeviceResponseTypes로 통일
  → battery_level이 number | null이 되어 실제 응답과 일치한다

폴링 주기는 통일하지 않는다. 온보딩(1초)은 기기가 붙는 순간 즉시 반응해야 하고,
설정(5초)은 상태 표시라 성격이 다르다.
