--- src/pages/onboarding/types/deviceTypes.ts
@@ -1,8 +1,0 @@
-export interface DeviceTypes {
-  id: number;
-  nickname: string;
-  mac_address: string;
-  battery_level: number;
-  is_connected: boolean;
-  last_seen_at: string;
-}
