+++ src/pages/setting/apis/patchDevice.ts
@@ -1,7 +1,7 @@
 import type {
   DeviceResponseTypes,
   UpdateDeviceRequestTypes,
-} from '@/pages/setting/types/deviceApiTypes';
+} from '@/shared/types/deviceApiTypes';
 import http from '@/shared/apis/axios';
 
 /** 기기 정보(이름) 수정. */
