+++ src/shared/types/deviceApiTypes.ts
